% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{del_genes}
\alias{del_genes}
\title{Delete genes from Seurat object with prefix}
\usage{
del_genes(seuratObj, gene_name_prefix, assay = "RNA")
}
\arguments{
\item{seuratObj}{Input Seurat object}

\item{assay}{Assay, defaults to RNA}

\item{gene_name}{Prefix of genes to filter out (delete); also takes vector of strings}
}
\value{
Seurat object with deleted genes
}
\description{
Based on \href{https://github.com/satijalab/seurat/issues/2610#issuecomment-585935810}{this GitHub issue.}
}
\examples{
seuratObj <- del_genes(seuratObj, "HLA")
# multiple genes
gene_name_prefixes <- c("HLA", "RPS", "RPL", "IG")
seuratObj <- del_genes(seuratObj, gene_name_prefixes, assay = "SCT")
}
