% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{doubFinder}
\alias{doubFinder}
\title{Predict and remove doublets from Seurat object}
\usage{
doubFinder(sample, sweep.stats.sample, rm.doubs = FALSE)
}
\arguments{
\item{sample}{Seurat object}

\item{sweep.stats.sample}{Summary stats from sum.sweep()}

\item{rm.doubs}{logical; default FALSE; removes doublets from \code{sample} if TRUE}
}
\value{
A Seurat object with identified doublets
}
\description{
Predict and remove doublets from Seurat object
}
\examples{

for (i in 1:length(gex_obj)) {
  message(crayon::red$underline$bold(paste0(
    "Finding Doublets with DoubletFinder for ", gex_names[[i]]
  )))
  gex_doubs[[i]] <-
    doubFinder(gex_obj[[i]], sweep.stats.sample = gex_pk[[i]])
  message(crayon::red$underline$bold(paste0("Saving ", gex_names[[i]])))
  save(gex_doubs, file = here("gex_obj_postDoublet_actual.RData"))
}
# with future
combined.dat <- future.apply::future_lapply(combined.dat, doubFinder, future.seed = TRUE)
}
