% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{fix_cellxgene_py}
\alias{fix_cellxgene_py}
\title{Fix h5ad to work with cellxgene}
\usage{
fix_cellxgene_py(file, file_out = NULL)
}
\arguments{
\item{file}{Path to h5ad converted from h5Seurat by SeuratDisk}

\item{file_out}{Output path, overwrites if blank}
}
\value{
In-place function, no return
}
\description{
For some reason, Seurat objects which have been converted to h5ad using SeuratDisk have some sort of indexing issue
which prevent them from being used with cellxgene or being rewritten after importing into Python. This function
uses Reticulate to call a few lines in Python to fix this issue. Learn more \href{https://github.com/theislab/scvelo/issues/255#issuecomment-739995301}{here}.
}
\examples{
SaveH5Seurat(
  samples.combined,
  filename = "processed/02-annotation/cellxgene_out_UNFILTERED_no4k_doublets.h5Seurat",
  verbose = TRUE,
  overwrite = TRUE
)

Convert(
  source = "processed/02-annotation/cellxgene_out_UNFILTERED_no4k_doublets.h5Seurat",
  dest = "h5ad",
  assay = "combatBatch",
  verbose = TRUE,
  overwrite = TRUE
)

fix_cellxgene(file = "processed/02-annotation/cellxgene_out_UNFILTERED_no4k_doublets.h5ad")
}
