% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{get_percent}
\alias{get_percent}
\title{Adds layers to Seurat object, percent mitochondrial and ribosomal genes}
\usage{
get_percent(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat object with RNA assay}
}
\value{
A Seurat object with two new layers
}
\description{
Matches string prefixes to gene symbols ("MT" and "RPS") to find percent mitochondrial and ribosomal genes per cell
}
\examples{
for (i in 1:length(gex_obj)) {
 gex_obj[[i]] <- get.percent(gex_obj[[i]])
}
}
