% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{load.seur.data}
\alias{load.seur.data}
\title{Load Cell Ranger count data}
\usage{
load.seur.data(dir, s, extra = "")
}
\arguments{
\item{dir}{root of CellRanger count output}

\item{s}{list of sample names/directories output from CellRanger count}

\item{extra}{file path modification if multiple CellRanger count instances exist}
}
\value{
Seurat object
}
\description{
Load Cell Ranger count data
}
\examples{
snames <- c('2hr_1', '2hr_2', '2hr_3', '2hr_ctrl', '2hr_untag', '12hr_1', '12hr_2', '12hr_3')
seq2.dat <- mclapply(snames, function(x) load.seur.data(x, "rerun"), mc.cores = 4)
snames <- paste(snames, "seq2", sep="_")
names(seq2.dat) <- snames
}
