% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{plot.mito}
\alias{plot.mito}
\title{Plots mitochondrial and ribosomal genes with ggplot2}
\usage{
\method{plot}{mito}(seuratObj, mito.cutoff = 10, rps.cutoff = 10, title = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object with percent mitochondrial and ribosomal genes calculated with get.percent()}

\item{mito.cutoff}{Cutoff of percent mitochondrial genes, default 10}

\item{rps.cutoff}{Cutoff of percent ribosomal genes, default 10}

\item{title}{Title of ggplot, default NULL}
}
\value{
A dual-faceted ggplot printed to active device
}
\description{
Plots mitochondrial and ribosomal genes with ggplot2
}
\examples{

plot.mito(gex_obj[[1]], title = gex_names[[1]])
plot.mito(gex_obj[[2]], title = gex_names[[2]])
plot.mito(gex_obj[[3]], title = gex_names[[3]])
plot.mito(gex_obj[[4]], title = gex_names[[4]])
plot.mito(gex_obj[[6]], title = gex_names[[6]])
plot.mito(gex_obj[[5]], title = gex_names[[5]])
}
