% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{plot_pk}
\alias{plot_pk}
\title{Plots result from find.pk}
\usage{
plot_pk(sweep.stats.sample, title = NA)
}
\arguments{
\item{sweep.stats.sample}{Summary stats generated by sum.sweep}

\item{title}{Plot title, default NA}
}
\value{
R plot printed to active device
}
\description{
Plots result from find.pk
}
\examples{

pdf(here("processed/01-5_3-comp/ParamSweepPeaks.pdf"))

plot.pk(gex_pk[[1]], title = gex_names[[1]])
plot.pk(gex_pk[[2]], title = gex_names[[2]])
plot.pk(gex_pk[[3]], title = gex_names[[3]])
plot.pk(gex_pk[[4]], title = gex_names[[4]])
plot.pk(gex_pk[[5]], title = gex_names[[5]])
plot.pk(gex_pk[[6]], title = gex_names[[6]])

dev.off()
}
