% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{prepGeneList}
\alias{prepGeneList}
\title{Prepare ranked gene list for GSEA}
\usage{
prepGeneList(DESeq_res, gene_counts, rank = "log2")
}
\arguments{
\item{DESeq_res}{DESeq2 Results}

\item{gene_counts}{Raw counts file}

\item{rank}{Rank by either "log2FC" or "padj"}
}
\value{
A named vector with gene names and either log2FC or padj in descending order
}
\description{
Prepare ranked gene list for GSEA
}
\examples{
geneList_res_WTvKO_cond1_input <- prepGeneList(res_WTvKO_cond1, gene_counts)

gsea_gene_WTvKO_cond1_sig_genes_tx <- gseGO(
   geneList     = geneList_res_WTvKO_cond1_input,
   OrgDb        = org.Mm.eg.db,
   keyType = "ENSEMBL",
   pAdjustMethod = "BH",
   ont = "BP",
   minGSSize    = 100,
   maxGSSize    = 500,
   pvalueCutoff = 0.10,
   verbose      = TRUE
)
}
