% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{process.combat}
\alias{process.combat}
\title{ComBat Batch-Correct SCT Counts in Seurat Object}
\usage{
process.combat(seuratObj, seuratBatch)
}
\arguments{
\item{seuratObj}{A Seurat object with an SCT assay}

\item{seuratBatch}{A Vector of batch labels for each}
}
\value{
Seurat object with combatBatch Assay
}
\description{
ComBat Batch-Correct SCT Counts in Seurat Object
}
\examples{
ComBat_merge <-
  merge(
    rna_seq[[1]],
    y = c(
      rna_seq[[2]],
      rna_seq[[3]],
      rna_seq[[4]],
      rna_seq[[5]],
      rna_seq[[6]],
      gex_doubs[[1]],
      gex_doubs[[2]],
      gex_doubs[[3]],
      gex_doubs[[4]],
      gex_doubs[[5]],
      gex_doubs[[6]]
    )
  )

combat_proccess <-
  process.combat(ComBat_merge, ComBat_merge$comp.ident)
}
