% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{remove_n_cells}
\alias{remove_n_cells}
\title{Remove Bottom or Top n Cells by nFeature_RNA or nCount_RNA}
\usage{
remove_n_cells(seuratObj, n, nFeature = TRUE, top = TRUE)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{n}{Number of desired cells to remove}

\item{nFeature}{Whether to sort by nFeature or nCount}

\item{top}{Top n cells by feature or bottom}
}
\value{
A Seurat object sorted by feature and with n cells removed
}
\description{
Remove Bottom or Top n Cells by nFeature_RNA or nCount_RNA
}
\examples{
# to remove the bottom 4k cells by nFeature_RNA:
seuratObj <- remove_n_cells(seuratObj, 4000, top = FALSE)
}
