% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{seurat_counts}
\alias{seurat_counts}
\title{Generate feature plots of transcript counts vs gene counts}
\usage{
seurat_counts(seuratObj, count.cutoff = 10000, downsample = 0)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{count.cutoff}{A cutoff indicating max counts per cell; default 10k}

\item{downsample}{A number representing number of cells to subset}
}
\value{
A FeatureScatter printed to device
}
\description{
Generate feature plots of transcript counts vs gene counts
}
\examples{

seurat.counts(gex_obj[[1]], count.cutoff = 22000) + ggtitle(gex_names[[1]])
seurat.counts(gex_obj[[2]], count.cutoff = 52000) + ggtitle(gex_names[[2]])
seurat.counts(gex_obj[[3]], count.cutoff = 10000) + ggtitle(gex_names[[3]])
seurat.counts(gex_obj[[4]], count.cutoff = 70000) + ggtitle(gex_names[[4]])
seurat.counts(gex_obj[[5]], count.cutoff = 65000) + ggtitle(gex_names[[5]])
seurat.counts(gex_obj[[6]], count.cutoff = 60000) + ggtitle(gex_names[[6]])

}
