% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{sum.sweep}
\alias{sum.sweep}
\title{Running paramSweep and summarizeSweep from DoubletFinder}
\usage{
\method{sum}{sweep}(sample, cores = 1)
}
\arguments{
\item{sample}{Seurat object}

\item{cores}{Number of cores, default 1}
}
\value{
Summary states from summarizeSweep
}
\description{
Running paramSweep and summarizeSweep from DoubletFinder
}
\examples{
for (i in 1:length(gex_obj)) {
  message(crayon::red$underline$bold(
  paste0("Finding Parameter Sweep with DoubletFinder for ", gex_names[[i]])
))
gex_pk[[i]] <-
  sum.sweep(gex_obj[[i]])
  message(crayon::red$underline$bold(paste0("Saving ", gex_names[[i]])))
  save(gex_pk, file = here("gex_obj_postDoublet.RData"))
}
}
