% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{viz.doubs}
\alias{viz.doubs}
\title{Visualize doublets on UMAP embedding w/ RNA counts for reference}
\usage{
viz.doubs(seuratObj, title = "Sample")
}
\arguments{
\item{seuratObj}{Seurat object with predicted doublets}

\item{title}{Sample name}
}
\value{
Printed ggplot2 to active device
}
\description{
Visualize doublets on UMAP embedding w/ RNA counts for reference
}
\examples{
pdf(
  here("processed/01-5_3-comp/GEX_doubletClusters.pdf"),
    height = 8,
    width = 14
  )

viz.doubs(gex_doubs[[1]], title = gex_names[[1]])
viz.doubs(gex_doubs[[2]], title = gex_names[[2]])
viz.doubs(gex_doubs[[3]], title = gex_names[[3]])
viz.doubs(gex_doubs[[4]], title = gex_names[[4]])
viz.doubs(gex_doubs[[5]], title = gex_names[[5]])
viz.doubs(gex_doubs[[6]], title = gex_names[[6]])

dev.off()
}
