% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{writeSeurat}
\alias{writeSeurat}
\title{Writes Seurat object to file}
\usage{
writeSeurat(seuratObj, assay = "RNA", dir_out = "./")
}
\arguments{
\item{seuratObj}{a Seurat object with UMAP coordinates}

\item{dir_out}{output directory, default cwd}

\item{assay_write}{Assay to write to file, default RNA}
}
\value{
Nothing, writes to file
}
\description{
A "manual" solution to writing a Seurat object to file
}
\examples{
writeSeurat(gex_obj[[i]], assay = "RNA")
}
