% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA-seq_fxns.R
\name{seurat.process}
\alias{seurat.process}
\title{Preprocess Seurat object}
\usage{
seurat.process(
  counts,
  counts_name,
  count.cutoff = 10000,
  mito.cutoff = 10,
  rps.cutoff = 10
)
}
\arguments{
\item{counts}{Seurat object}

\item{counts_name}{Name of sample corresponding to seurat object}

\item{count.cutoff}{Cutoff of counts used in seurat.counts, default 10k}

\item{mito.cutoff}{Cutoff of percent mitochondrial genes per cell, default 10}

\item{rps.cutoff}{Cutoff of percent ribosomal genes per cell, default 10}
}
\value{
Seurat object that has been QC'd, SCTransformed, PCA'd, Clustered and UMAP.
}
\description{
Preprocess Seurat object
}
\examples{
for (i in 1:length(gex_obj)) {
gex_obj[[i]] <-
  seurat.process(gex_obj[[i]], counts_name = gex_names[[i]])
}
}
